c'est le bordel aled

+++ FourmisRPOOIDL/src/fourmilliere/Reserve.java
@@ -1,28 +1,28 @@
 package fourmilliere;
 
 public class Reserve {
-	private long ReserveNouriture=0;
+	private double ReserveNouriture=0.0;
 	Fourmilliere fourmilliere=null;
 	
 	public Reserve(Fourmilliere fourmilliere){
 		this.fourmilliere=fourmilliere;
 	}
 	
-	public void AjouterNourriture(int nourriture){
+	public void AjouterNourriture(double nourriture){
 		this.ReserveNouriture+=nourriture;
 	}
 	public boolean PrendreNourriture(double nourriture){
-		if(this.ReserveNouriture-nourriture>=0)
+		if(this.ReserveNouriture-nourriture>=0.0)
 			this.ReserveNouriture-=nourriture;
 		else
 			return false;
 		return true;
 	}
 	public boolean EstVide(){
-		if(this.ReserveNouriture==0) return true;
+		if(this.ReserveNouriture==0.0) return true;
 		return false;
 	}
-	public long getReserveNouriture(){
+	public double getReserveNouriture(){
 		return this.ReserveNouriture;
 	}
 }
